requirements.txt
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0

# Phase 2 (필요 시 설치)
# psd-tools>=1.9.0

# Phase 2 (SAM 백엔드)
# segment-anything

# Phase 3 (Grounded SAM 백엔드)
# groundingdino-py
# transformers>=4.35.0
